minor improvements - fix

+++ src/components/mac-rent-info-row/index.js
@@ -2,10 +2,18 @@ import React, {Component} from "react";
 import {Popover, OverlayTrigger} from "react-bootstrap";
 import {withRouter} from "react-router-dom";
 import PropTypes from "prop-types";
-
-var Button = require('antd/lib/button');
+import moment from "moment";
+import {confirmAlert} from "react-confirm-alert";
+import "react-confirm-alert/src/react-confirm-alert.css";
+import { Button } from "antd";
 
  export default withRouter( class MacRentInfoRow extends Component {
+   constructor () {
+        super();
+        this.state= {
+            todayDate: moment().format("DD/MM/YYYY")
+        }
+    };
 
     static propTypes = {
         id: PropTypes.number,
@@ -25,9 +33,16 @@ var Button = require('antd/lib/button');
         this.props.history.push(`/input/${this.props.userName}-${this.props.id}`);
     }
     handleDeleteButton(){
-        fetch(`http://localhost:3456/mac-rent-informations/${this.props.id}`, {
-            method: "DELETE"
-        }).then(response => response.json().then(json => {return json;})); }
+        confirmAlert({
+            title: "Eliminare?",
+            message: "",
+            confirmLabel: "OK",
+            cancelLabel: "Annulla",
+            onConfirm: () => fetch(`http://localhost:3456/mac-rent-informations/${this.props.id}`, {
+                method: "DELETE"
+            }).then(response => response.json().then(json => {return json;})),
+        })
+    }
 
     render () {
         const {
@@ -42,13 +57,17 @@ var Button = require('antd/lib/button');
             owner,
             serial
         } = this.props
+        var b = moment(dateFrom, "dd/mm/yyyy");
+        var c = moment(dateTo, "dd/mm/yyyy")
+        var a = moment(moment().format("DD/MM/YYYY"), "dd/mm/yyyy");
+
         return (
             <tr>
                 <th>{id}</th>
                 <th>{name}</th>
                 <th>{code}</th>
-                <th>{dateFrom}</th>
-                <th>{dateTo}</th>
+                <th style={{backgroundColor:(moment(b).isBefore(a)? "": "#99ff99")}}>{dateFrom}</th>
+                <th style={{backgroundColor:(moment(c).isBefore(a)? "#ff5555": "")}}>{dateTo}</th>
                 <th>{serial}</th>
                 <th>{owner}</th>
                 <th>{fee} €</th>
